Added GUI to UserInterface class using Java Swing
The User Interface class now uses Java Swing as a GUI instead of command prompt. The new interface also ensures the user gives all information necessary for the creation of User Request.

Java Swing was added instead of JavaFX because it is easier to manage dependencies and we do not need an updated GUI framework for a very simple GUI